Audiracmichelle/issue5

+++ src/create_data_symlinks.py
@@ -1,5 +1,8 @@
+import logging
 import os
 import hydra
+
+LOGGER = logging.getLogger(__name__)
 
 @hydra.main(config_path="conf", config_name="config", version_base=None)
 def main(cfg):
@@ -10,11 +13,15 @@ def main(cfg):
             
             subfolder_path = f"data/{datapath}/{subfolder_name}"
 
+            if os.path.exists(subfolder_path):
+                LOGGER.info(f"Subfolder {subfolder_path} already exists.")
+                continue
             if subfolder_link is not None:
                 os.symlink(subfolder_link, subfolder_path)
-            
+                LOGGER.info(f"Created symlink {subfolder_path} -> {subfolder_link}")
             else:
                 os.makedirs(subfolder_path, exist_ok=True)
+                LOGGER.info(f"Created subfolder {subfolder_path}")
 
 if __name__ == "__main__":
     main()
